Rework from scratch
It was a drastic move, but I'm quite happy with the results. Some of the
state management is still a little hairy, but I think it is clear enough
to get us through. I am almost done with the editing aspects of the UI,
and I'm now confident that I can finish this up.