Allow mouse clicks to propagate
Previously, I had stopped any "mouseUp" events on nodes and control
points from propagating to the canvas. This was in order to prevent
de-selection when the user clicked a node. However, it had less
desirable side effects (for instance, when the user clicked a node while
a text input was focused, the text input was not blurred as one would
expect). I've changed things to allow these events to propagate.