Reverse default control point "bias"
In order to render the transition details in an aesthetically pleasing
way, the control point needs to be displaced from the line joining the
two endpoints of its associated arrow. When initially creating a new
control point, we offset the control point slightly. I've simply
reversed the direction of this offset in order to match what I *presume*
is the more natural direction to add arrows.