Fix/trade bug: wonky behavior when moving node
Previously, if a user focused a node's mnemonic and then attempted to
move the node, only associated control points would move. This was due
to the fact that the "blur" event for the mnemonic was fired AFTER the
mouse events related to the node, and the blur event resulted in the WIP
state being nullified. I've "fixed" this by simply preserving the WIP
state when the mnemonic is blurred, although this merely shifts the
problem: now any changes to the mnemonic aren't saved if the user clicks
a node while the mnemonic is focused. However, it is a better tradeoff.